refactor(components): refactor carousel / carousel-item
