+++ packages/components/carousel/index.ts
@@ -1,6 +1,6 @@
 import { withInstall, withNoopInstall } from '@element-plus/utils/with-install'
-import Carousel from './src/main.vue'
-import CarouselItem from './src/item.vue'
+import Carousel from './src/carousel.vue'
+import CarouselItem from './src/carousel-item.vue'
 
 export const ElCarousel = withInstall(Carousel, {
   CarouselItem,
