+++ packages/components/carousel/src/carousel.ts
@@ -1,4 +1,6 @@
-import type { Ref, ToRefs, UnwrapRef } from 'vue'
+import { buildProps } from '@element-plus/utils/props'
+import { isNumber } from '@element-plus/utils/util'
+import type { Ref, ToRefs, UnwrapRef, ExtractPropTypes } from 'vue'
 
 export interface ICarouselItemProps {
   name: string
@@ -49,3 +51,55 @@ export interface InjectCarouselScope {
   removeItem: (uid: number) => void
   setActiveItem: (index: number) => void
 }
+
+export const carouselProps = buildProps({
+  initialIndex: {
+    type: Number,
+    default: 0,
+  },
+  height: { type: String, default: '' },
+  trigger: {
+    type: String,
+    default: 'hover',
+  },
+  autoplay: {
+    type: Boolean,
+    default: true,
+  },
+  interval: {
+    type: Number,
+    default: 3000,
+  },
+  indicatorPosition: { type: String, default: '' },
+  indicator: {
+    type: Boolean,
+    default: true,
+  },
+  arrow: {
+    type: String,
+    default: 'hover',
+  },
+  type: { type: String, default: '' },
+  loop: {
+    type: Boolean,
+    default: true,
+  },
+  direction: {
+    type: String,
+    default: 'horizontal',
+    values: ['horizontal', 'vertical'],
+  },
+  pauseOnHover: {
+    type: Boolean,
+    default: true,
+  },
+} as const)
+
+export type CarouselProps = ExtractPropTypes<typeof carouselProps>
+
+export const carouselEmits = {
+  change: (current: number, prev: number) =>
+    isNumber(current) && isNumber(prev),
+}
+
+export type CarouselEmits = typeof carouselEmits
